pyproject.toml
[project]
name = "momir-printer"
version = "0.1.0"
description = "Momir format MTG creature printer for Phomemo M02S"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "bleak>=0.22.0",
    "Pillow>=11.0.0",
    "httpx>=0.28.0",
    "pymupdf>=1.25.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
